💥 Drop Nicercast in favour of express and spawn
Migration guide:

PulseAudio have been dropped in favour of ALSA, don't install pulseaudio on the Raspberry PI.

This program now operates in two modes; "passthrough" and "source". In "passthrough" mode the audio is passed from line in to line out, and the physical keys control the volume. In "source" mode the sound is streamed over TCP, and the physical keys doesn't do anything.